[refactor] organize each code by file
