+++ feature/src/main/java/com/begin_a_gain/feature/match/common/MatchCategory.kt
@@ -3,14 +3,22 @@ package com.begin_a_gain.feature.match.common
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.layout.wrapContentWidth
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.SheetState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableIntStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.layout.Layout
@@ -18,6 +26,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import com.begin_a_gain.feature.match.create_match.util.constant.categories
+import com.begin_a_gain.library.design.component.bottom_sheet.OBottomSheet
+import com.begin_a_gain.library.design.component.button.OButton
 import com.begin_a_gain.library.design.component.text.OText
 import com.begin_a_gain.library.design.theme.ColorToken
 import com.begin_a_gain.library.design.theme.ColorToken.Companion.color
@@ -107,5 +117,38 @@ fun CategoryChip(
             color = if (isSelected) ColorToken.TEXT_PRIMARY
             else ColorToken.TEXT_01
         )
+    }
+}
+
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun CategoryBottomSheet(
+    sheetState: SheetState,
+    selectedIndex: Int,
+    onDismissRequest: () -> Unit,
+    onSelected: (Int) -> Unit
+) {
+    var currentIndex by rememberSaveable { mutableIntStateOf(selectedIndex) }
+
+    OBottomSheet(
+        title = "대국 카테고리",
+        sheetState = sheetState,
+        onDismissRequest = onDismissRequest
+    ) {
+        Column(
+            modifier = Modifier.padding(horizontal = 20.dp, vertical = 16.dp)
+        ) {
+            MatchCategoryGrid(selectedIndex) {
+                currentIndex = it
+            }
+            Spacer(modifier = Modifier.weight(1f))
+            OButton(
+                modifier = Modifier.fillMaxWidth(),
+                text = "확인"
+            ) {
+                onSelected(currentIndex)
+            }
+        }
     }
 }
