+++ feature/src/main/java/com/begin_a_gain/feature/match/create_match/CreateMatchScreen.kt
@@ -1,63 +1,42 @@
 package com.begin_a_gain.feature.match.create_match
 
-import android.util.Log
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.expandIn
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.animation.shrinkOut
-import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.SheetState
 import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.focus.FocusDirection
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.focus.onFocusChanged
-import androidx.compose.ui.platform.LocalFocusManager
-import androidx.compose.ui.platform.LocalSoftwareKeyboardController
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.begin_a_gain.feature.match.common.MatchCategoryGrid
+import com.begin_a_gain.feature.match.common.CategoryBottomSheet
 import com.begin_a_gain.feature.match.create_match.util.type.RepeatDayType
-import com.begin_a_gain.feature.match.create_match.util.ui.CodeTextField
+import com.begin_a_gain.feature.match.create_match.util.ui.DaySelection
 import com.begin_a_gain.feature.match.create_match.util.ui.MatchCodeDialog
 import com.begin_a_gain.library.design.component.ODivider
-import com.begin_a_gain.library.design.component.bottom_sheet.OBottomSheet
 import com.begin_a_gain.library.design.component.bottom_sheet.OPickerBottomSheet
 import com.begin_a_gain.library.design.component.button.ButtonType
-import com.begin_a_gain.library.design.component.button.OButton
-import com.begin_a_gain.library.design.component.dialog.ODialog
 import com.begin_a_gain.library.design.component.image.OImage
 import com.begin_a_gain.library.design.component.image.OImageRes
 import com.begin_a_gain.library.design.component.selection.OSwitch
@@ -67,7 +46,6 @@ import com.begin_a_gain.library.design.theme.ColorToken
 import com.begin_a_gain.library.design.theme.ColorToken.Companion.color
 import com.begin_a_gain.library.design.theme.OTextStyle
 import com.begin_a_gain.library.design.util.OScreen
-import com.begin_a_gain.library.design.util.noRippleClickable
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Preview
@@ -220,38 +198,6 @@ fun CreateMatchScreen(
     }
 }
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun CategoryBottomSheet(
-    sheetState: SheetState,
-    selectedIndex: Int,
-    onDismissRequest: () -> Unit,
-    onSelected: (Int) -> Unit
-) {
-    var currentIndex by rememberSaveable { mutableIntStateOf(selectedIndex) }
-
-    OBottomSheet(
-        title = "대국 카테고리",
-        sheetState = sheetState,
-        onDismissRequest = onDismissRequest
-    ) {
-        Column(
-            modifier = Modifier.padding(horizontal = 20.dp, vertical = 16.dp)
-        ) {
-            MatchCategoryGrid(selectedIndex) {
-                currentIndex = it
-            }
-            Spacer(modifier = Modifier.weight(1f))
-            OButton(
-                modifier = Modifier.fillMaxWidth(),
-                text = "확인"
-            ) {
-                onSelected(currentIndex)
-            }
-        }
-    }
-}
-
 @Composable
 fun SettingBox(
     label: String,
@@ -316,50 +262,6 @@ fun SettingRow(
                 onCheckedChanged()
             }
             Spacer(modifier = Modifier.width(16.dp))
-        }
-    }
-}
-
-@Preview
-@Composable
-fun DaySelection(
-    selectedDays: List<Boolean> = (1..7).map { false },
-    onClick: (Int) -> Unit = {}
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 20.dp)
-            .padding(bottom = 16.dp)
-    ) {
-        listOf("일", "월", "화", "수", "목", "금", "토").forEachIndexed { index, day ->
-            val selected = selectedDays[index]
-            Box(
-                modifier = Modifier
-                    .size(38.dp)
-                    .clip(CircleShape)
-                    .border(
-                        width = 1.dp,
-                        shape = CircleShape,
-                        color = if (selected) ColorToken.STROKE_PRIMARY.color() else ColorToken.STROKE_02.color()
-                    )
-                    .background(
-                        if (selected) ColorToken.UI_PRIMARY.color() else ColorToken.UI_BG.color()
-                    )
-                    .clickable {
-                        onClick(index)
-                    },
-                contentAlignment = Alignment.Center
-            ) {
-                OText(
-                    text = day,
-                    style = OTextStyle.Subtitle2,
-                    color = if (selected) ColorToken.TEXT_ON_01 else ColorToken.TEXT_02
-                )
-            }
-            if (index < 6) {
-                Spacer(modifier = Modifier.weight(1f))
-            }
         }
     }
 }
