+++ feature/src/main/java/com/begin_a_gain/feature/match/create_match/util/ui/CodeComponent.kt
@@ -19,7 +19,6 @@ import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.begin_a_gain.library.design.component.dialog.ODialog
 import com.begin_a_gain.library.design.component.text.OTextField
